Remove last paragraphs of III.C
This wipes the incomplete paragraphs introduced in 5aa6a4d.

The deleted paragraphs had attempted to explain an unphysical
phenomenon present in the analytical solutions (unphysical because
no scattering is allowed), and reproduced robustly in the numerical
solutions. The phenomenon isn't visible in any of the plots
presented with this test, however; and furthermore, an analysis of
the phenomenon doesn't lead to anything else in the paper. So I
deleted it, rather than labor any longer over improving the
paragraphs.